examples/virtio: converge with examples/virtio-socket
