--- examples/virtio-socket/client_vm/gic_v2_overlay.dts
@@ -1,19 +1,0 @@
-/*
- * Copyright 2025, UNSW
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-/ {
-	intc@8000000 {
-		phandle = <0x8001>;
-		interrupts = <0x01 0x09 0x04>;
-		reg = <0x00 0x8000000 0x00 0x10000 0x00 0x8010000 0x00 0x10000 0x00 0x8030000 0x00 0x10000 0x00 0x8040000 0x00 0x10000>;
-		compatible = "arm,cortex-a15-gic";
-		ranges;
-		#size-cells = <0x02>;
-		#address-cells = <0x02>;
-		interrupt-controller;
-		#interrupt-cells = <0x03>;
-	};
-};
