--- examples/virtio-socket/client_vm_userlevel/recver_init.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-
-# Copyright 2025, UNSW
-#
-# SPDX-License-Identifier: BSD-2-Clause
-
-CID=3
-/root/vsock_recv.elf "$CID"
