+++ examples/virtio/client_vm/linux.dts
@@ -66,4 +66,10 @@
 		reg = <0x00 0x160000 0x00 0x200>;
 		interrupts = <0x00 44 0x04>;
 	};
+
+	virtio-socket@170000 {
+		compatible = "virtio,mmio";
+		reg = <0x00 0x170000 0x00 0x200>;
+		interrupts = <0x00 45 0x04>;
+	};
 };
