+++ examples/virtio/client_vm/linux_config
@@ -1090,7 +1090,11 @@ CONFIG_DEFAULT_TCP_CONG="cubic"
 CONFIG_DNS_RESOLVER=y
 # CONFIG_BATMAN_ADV is not set
 # CONFIG_OPENVSWITCH is not set
-# CONFIG_VSOCKETS is not set
+CONFIG_VSOCKETS=y
+CONFIG_VSOCKETS_DIAG=y
+CONFIG_VSOCKETS_LOOPBACK=y
+CONFIG_VIRTIO_VSOCKETS=y
+CONFIG_VIRTIO_VSOCKETS_COMMON=y
 # CONFIG_NETLINK_DIAG is not set
 # CONFIG_MPLS is not set
 # CONFIG_NET_NSH is not set
@@ -2299,6 +2303,7 @@ CONFIG_VIRTIO_MMIO=y
 # CONFIG_VDPA is not set
 CONFIG_VHOST_MENU=y
 # CONFIG_VHOST_NET is not set
+# CONFIG_VHOST_VSOCK is not set
 # CONFIG_VHOST_CROSS_ENDIAN_LEGACY is not set
 
 #
